Import all Fedora patches
Import all of them, but comment out a few in the spec file.
Specifically:
- patches related to .gitignore that is not part of upstream tarball
- some patches for tools that are irrelevant for just the bootloader
  build
- systemd units irrelevant for PV(H) grub on dom0 side

There are a lot more that are not relevant for the PV(H)Grub case (all
for architectures not currently supported like sparc or powerpc,
several patches for network boot and more). But they should also
(hopefully) be harmless.

And also update build dependencies based on Fedora package.

Fixes QubesOS/qubes-issues#9471